requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
pushbullet.py==0.11.0
python-magic==0.4.13
PyYAML==3.12
requests==2.18.1
six==1.10.0
urllib3==1.21.1
websocket-client==0.44.0
